+++ test/typescript/t.test.ts
@@ -132,6 +132,21 @@ function interpolation(t: TFunction) {
   t('arrayJoin', { joinArrays: '+' });
   // -> "line1+line2+line3"
 
+  const resolved = t('key', { returnDetails: true });
+  resolved.res;
+  resolved.res.substring(2, 1);
+  resolved.usedKey;
+  resolved.exactUsedKey;
+  resolved.usedNS;
+  resolved.usedLng;
+
+  const r2 = t('keyTwo', { returnDetails: false });
+  r2.substring(0, 2); // make sure it is a string
+  const r3 = t('keyThree');
+  r3.substring(0, 2); // make sure it is a string
+  const r4 = t('keyTwo', { ns: 'whatever' });
+  r4.substring(0, 2); // make sure it is a string
+
   t('arrayJoinWithInterpolation', {
     myVar: 'interpolate',
     joinArrays: ' ',